+++ frontends/web/entries/react.tsx
@@ -1,0 +1,8 @@
+export {
+    Boytacean,
+    type BoytaceanHandle,
+    type BoytaceanProps
+} from "../react/components/boytacean/boytacean";
+export { BACKGROUNDS, EMBED_STORAGE_PREFIX } from "../ts/embed";
+export { GameboyEmulator, SerialDevice } from "../ts/gb";
+export { PALETTES, PALETTES_MAP } from "../ts/palettes";